Actualización con Diprocom QA

+++ DIPROCOM/Tables/SWIFT_DOCUMENT_SEQUENCE.sql
@@ -4,16 +4,18 @@
     [ASSIGNED_DATETIME]    DATETIME      NULL,
     [POST_DATETIME]        DATETIME      NULL,
     [ASSIGNED_BY]          VARCHAR (100) NULL,
-    [DOC_FROM]             INT           NULL,
-    [DOC_TO]               INT           NULL,
+    [DOC_FROM]             BIGINT        NULL,
+    [DOC_TO]               BIGINT        NULL,
     [SERIE]                VARCHAR (100) NOT NULL,
     [ASSIGNED_TO]          VARCHAR (100) NULL,
-    [CURRENT_DOC]          INT           NULL,
+    [CURRENT_DOC]          BIGINT        NULL,
     [STATUS]               VARCHAR (15)  NULL,
     [BRANCH_NAME]          VARCHAR (50)  NULL,
     [BRANCH_ADDRESS]       VARCHAR (150) NULL,
     CONSTRAINT [PK_SWIFT_DOCUMENT_SEQUENCE] PRIMARY KEY CLUSTERED ([ID_DOCUMENT_SECUENCE] ASC)
 );
+
+
 
 
 GO
